Add files via upload
The data is now more valid and makes sure the user can not post or search if any of the fields are empty. The user can not click on Communities if they are in a community. They have to use their user name to log in to the app(hard coded only to check if it is empty).  Has a way to simulate a database and makes each  "offer/ride" clickable. it does not have a way to send the data so it simulates it.